fix: widen graphics IDs in source schema

+++ Persistence/database/npcs/structure.sql
@@ -12,7 +12,7 @@ CREATE TABLE `npcs`
     `shield_absorption` tinyint      NOT NULL DEFAULT 0.0,
     `damage`            int          NOT NULL DEFAULT 0,
     `speed`             smallint     NOT NULL DEFAULT 300,
-    `gfx`               tinyint      NOT NULL DEFAULT 0,
+    `gfx`               smallint     NOT NULL DEFAULT 0,
     `ship_type`         varchar(255) NOT NULL DEFAULT '',
     `ai`                tinyint      NOT NULL DEFAULT 1,
